Made JS Journey run on Phantom and close gracefully
