+++ package.json
@@ -7,14 +7,16 @@
     "build": "webpack",
     "test": "npm run test:unit",
     "start": "npm run build && node server",
-    "test:unit": "node test/.mocha.js",
-    "test:journey": "selenium-standalone start | serve ./public -p 3999 | node ./test/.wd-sync.js",
-    "test:journey-js": "serve ./public -p 3999 | node ./test/.selenium.js",
+    "test:unit": "node test/unit.js",
+    "test:journey": "selenium-standalone start | node server | node ./test/journey.js",
+    "test:journey-js": "npm run build && node ./test/journey-js.js",
+    "test:all": "selenium-standalone start | npm run start | node test/.mocha.js | node ./test/journey.js | node ./test/.selenium.js",
     "watch:test:unit": "nodemon -x 'npm run test:unit'",
-    "watch:test:journey": "selenium-standalone start | serve ./public -p 3999 | nodemon -x 'node ./test/.wd-sync.js'",
-    "watch:test:journey-js": "serve ./public -p 3999 | nodemon -x 'node ./test/.selenium.js'",
+    "watch:test:journey": "selenium-standalone start | npm run start | nodemon -x 'node ./test/journey.js'",
+    "watch:test:journey-js": "npm run start | nodemon -x 'node ./test/journey-js.js'",
+    "watch:test:all": "selenium-standalone start | npm run start | nodemon -x 'node test/.mocha.js | node ./test/journey.js | node ./test/journey-js.js'",
     "watch:serve": "webpack-dev-server --watch",
-    "git:push": "npm run test && git push",
+    "git:push": "npm run test:all && git push",
     "postinstall": "npm run test"
   },
   "engineStrict": true,
@@ -39,6 +41,7 @@
     "babel-register": "^6.4.3",
     "body-parser": "^1.15.2",
     "chai": "^3.5.0",
+    "chai-webdriver-promised": "^4.0.3",
     "css-loader": "0.25.0",
     "css-modules-require-hook": "^4.0.5",
     "dotenv": "^4.0.0",
@@ -51,6 +54,7 @@
     "mocha-ci-slack-reporter": "^1.0.1",
     "node-sass": "^3.10.0",
     "nodemon": "^1.11.0",
+    "phantomjs-prebuilt": "^2.1.14",
     "postcss-loader": "1.1.0",
     "react-addons-test-utils": "^15.4.0",
     "require-hacker": "^2.1.4",
@@ -58,7 +62,6 @@
     "sass-loader": "4.0.2",
     "selenium-html-js-converter": "^0.4.7",
     "selenium-webdriver": "^3.0.1",
-    "serve": "^1.4.0",
     "sinon": "^1.17.7",
     "style-loader": "0.13.1",
     "url-loader": "0.5.7",
