+++ server.js
@@ -16,9 +16,11 @@ app.get('/', function response(req, res) {
   res.sendFile(path.join(__dirname, 'public/index.html'));
 });
 
-app.listen(port, '0.0.0.0', function onStart(err) {
+var server = app.listen(port, '0.0.0.0', function onStart(err) {
   if (err) {
     console.log(err);
   }
   console.info('==> 🌎 Listening on port %s. Open up http://0.0.0.0:%s/ in your browser.', port, port);
 });
+
+ module.exports = server;
